refactor(routes): extract queue history router
